Front+Back v1/requirements.txt
Flask==3.0.0
requests==2.*
google-genai==1.11.0